fix(shokunin): minor tweaks on docs
