+++ src/loggers.rs
@@ -211,32 +211,6 @@ impl Default for Log {
         }
     }
 }
-
-/// Convenience macro to simplify log creation and logging.
-#[macro_export]
-macro_rules! macro_log_info {
-    ($level:expr, $component:expr, $description:expr, $format:expr) => {
-        {
-            use $crate::loggers::{Log, LogLevel, LogFormat};
-
-            extern crate dtt;
-            use dtt::DateTime;
-            // Get the current date and time in ISO 8601 format.
-            let date = DateTime::new();
-            let iso = date.iso_8601;
-
-            extern crate vrd;
-            use vrd::Random;
-            // Create a new random number generator
-            let mut rng = Random::default();
-            let session_id = rng.rand().to_string();
-
-            let log = Log::new(&session_id, &iso, $level, $component, $description, $format);
-            let _ = log.log();
-        }
-    };
-}
-
 #[cfg(test)]
 /// Tests for the `log_info!` macro.
 mod tests {
